requirements.txt
aiofiles>=23.2.1
httpx>=0.27.0
jinja2>=3.1.4
markdown>=3.6
Pillow>=10.3.0
playwright>=1.44.0
pydantic>=1.10.0,<2.0.0